requirements.txt
requests
bertopic
python-dotenv
umap-learn
torch
transformers